Added the Daemonite material implemetation
