+++ client/lib/material/dropdown.js
@@ -1,0 +1,13 @@
+// dropdown menu max-width
+	$('.dropdown').on('show.bs.dropdown', function () {
+		var $dropdownMenu = $('.dropdown-menu', $(this)),
+		    $dropdownToggle = $('[class*="dropdown-toggle"]', $(this)),
+		    dropdownPadding = $('a', $dropdownMenu).css('padding-left').replace('px', ''),
+		    dropdownWidth;
+		if ($dropdownMenu.hasClass('dropdown-menu-right') || $dropdownMenu.parents('.nav.pull-right').length) {
+			dropdownWidth = $dropdownToggle.offset().left + $dropdownToggle.outerWidth() - dropdownPadding;
+		} else {
+			dropdownWidth = $(window).width() - $dropdownToggle.offset().left - dropdownPadding;
+		}
+		$dropdownMenu.css('max-width', dropdownWidth);
+	});
