+++ client/lib/material/footer.js
@@ -1,0 +1,4 @@
+// footer push
+	function footerPush() {
+		$('body').css('margin-bottom', $('.footer').outerHeight());
+	}
