+++ client/lib/material/winresize.js
@@ -1,0 +1,11 @@
+// window resize
+	on_resize(function() {
+		// fixed left/right hand side column padding bottom and width
+			if ($('.content-fix').length) {
+				contentFixPushCal();
+			};
+		// footer push
+			footerPush();
+		// header height
+			headerHeightCal();
+	})();
